using uint64 for content IDs even though uint16 would be fine it's just easier to deal with I added a comment for voxels to indicate that 2^16 is the max
also, can now reference content by name not just id

+++ content/entity.go
@@ -2,7 +2,7 @@ package content
 
 import "github.com/felzix/huyilla/types"
 
-var EntityDefinitions = map[uint32]*types.EntityDefinition {
+var EntityDefinitions = map[uint64]*types.EntityDefinition {
     0: {Name: "human",
         Falls: true,
         InventoryCapacity: 10},
